Split local data into its own file
